+++ sandbox/src/pages/Buttons.tsx
@@ -1,7 +1,7 @@
 /* @license Copyright 2023 @paritytech/polkadot-cloud authors & contributors
 SPDX-License-Identifier: GPL-3.0-only */
 
-import { Button } from "packages/core-ui/lib/buttons";
+import { Button } from "packages/core-ui/lib/buttons/Button";
 import { faUser } from "@fortawesome/free-solid-svg-icons";
 import {
   faArrowAltCircleUp,
